update redis token

+++ config/default.conf.go
@@ -11,10 +11,10 @@ import (
 
 // server 服务基本配置结构
 type server struct {
-	RunMode   string `mapstructure:"runMode"`
-	Port      int    `mapstructure:"port"`
-	JWTSecret string `mapstructure:"jwtSecret"`
-	JWTExpire int64  `mapstructure:"jwtExpire"`
+	RunMode		string	`mapstructure:"runMode"`
+	Port		int		`mapstructure:"port"`
+	JWTSecret	string	`mapstructure:"jwtSecret"`
+	JWTExpire	int		`mapstructure:"jwtExpire"`
 }
 
 // ServerConf 服务基本配置
